Refactor: move bookmark models to bookmark/models
