+++ app/src/main/java/fr/free/nrw/commons/bookmarks/pictures/BookmarkPicturesController.java
@@ -1,7 +1,7 @@
 package fr.free.nrw.commons.bookmarks.pictures;
 
 import fr.free.nrw.commons.Media;
-import fr.free.nrw.commons.data.models.Bookmark;
+import fr.free.nrw.commons.bookmarks.models.Bookmark;
 import fr.free.nrw.commons.media.MediaClient;
 import io.reactivex.Observable;
 import io.reactivex.ObservableSource;
