+++ app/src/test/kotlin/fr/free/nrw/commons/bookmarks/pictures/BookmarkPictureDaoTest.kt
@@ -9,7 +9,7 @@ import android.net.Uri
 import android.os.RemoteException
 import com.nhaarman.mockitokotlin2.*
 import fr.free.nrw.commons.TestCommonsApplication
-import fr.free.nrw.commons.data.models.Bookmark
+import fr.free.nrw.commons.bookmarks.models.Bookmark
 import fr.free.nrw.commons.bookmarks.pictures.BookmarkPicturesContentProvider.BASE_URI
 import org.junit.runner.RunWith
 import org.robolectric.RobolectricTestRunner
